fix: move /types /shared
