+++ src/gas/controllers/Club/JoinController.ts
@@ -1,5 +1,5 @@
 import JoinModel from "../../models/Club/JoinModel";
-import JoinInterface from "../../types/JoinInterface";
+import JoinInterface from "../../shared/types/JoinInterface";
 
 export default class JoinController {
   private model = new JoinModel();
