+++ src/gas/controllers/Club/RegistController.ts
@@ -1,5 +1,5 @@
 import RegistModel from "../../models/Club/RegistModel";
-import RegistInterface from "../../types/RegistInterface";
+import RegistInterface from "../../shared/types/RegistInterface";
 
 export default class RegistController {
   private model = new RegistModel();
