+++ src/gas/controllers/InvalidActionController.ts
@@ -1,4 +1,4 @@
-import ResponseInterface from "../types/ResponseInterface";
+import ResponseInterface from "../shared/types/ResponseInterface";
 import InvalidActionView from "../views/InvalidActionView";
 
 export default class InvalidActionController {
