+++ src/gas/shared/types/ParameterInterface.ts
