+++ src/gas/shared/types/RegistInterface.ts
