+++ src/gas/shared/types/ResponseInterface.ts
