+++ src/gas/views/Club/JoinView.ts
@@ -1,5 +1,5 @@
 import Response from "../../shared/Response";
-import ResponseInterface from "../../types/ResponseInterface";
+import ResponseInterface from "../../shared/types/ResponseInterface";
 
 export default class JoinView {
   private res = new Response();
